changed default behaviour of tokenbutton click to only apply first property
